volunteer/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
httpx
groq
python-dotenv
apscheduler
pydantic